Revised module eval code for register updates
A subtle issue with register updates was fixed. A module's eval() function can be called multiple times during a clock cycle. However, in any one eval(), registers might be speculatively updated, but in later evals() not updated. However, the "source" state for the first eval() will be changed and not restored. This leads to potential surprises for users. The change here is to always restore all module registers to their replica state before any new eval().